requirements.txt
# Modules ML principaux
transformers~=4.50.0
torch~=2.7.0
accelerate~=0.26.0
sentencepiece
datasets~=3.6.0

# Utilitaires
gitpython~=3.1.44
scikit-learn
tqdm~=4.67.1
numpy
pandas
psutil~=7.0.0
requests~=2.32.3